float add pour class panier

+++ app/Controllers/shoppingCartController.php
@@ -24,7 +24,7 @@ class shoppingCartController extends Controller
             if (isset($_SESSION['quantite'])) {
                 // assignation valeur
                 $id_article = (int) $_POST['id_article'];
-                $prix_article = (int) $_POST['prix_article'];
+                $prix_article = (float) $_POST['prix_article'];
 
                 $_SESSION['quantite'][$id_article] = 1;
                 $_SESSION['prix'][$id_article] = $prix_article;
@@ -35,7 +35,7 @@ class shoppingCartController extends Controller
 
                 // assignation valeur
 
-                $prix_article = (int) $_POST['prix_article'];
+                $prix_article = (float) $_POST['prix_article'];
                 $id_article = (int) $_POST['id_article'];
                 $_SESSION['quantite'][$id_article] = 1;
                 $_SESSION['prix'][$id_article] = $prix_article;
